Complete list of available adl official color palettes

A list of all color palettes to choose from, and preview directly in the console.

Usage

adl_palettes

Format

A list of 9 color objects elements.

primary

Primary categorical color

secondary

Seconday categorical color

tertiary

Tertiary categorical color

pid3

Palette for partisanship/ideology (categorical)

categorical

Palette for categorical data for up to 7 groups (categorical)

likert_6

Palette for likert scales with 6 response options (diverging)

likert_4

Palette for likert scales with 4 response options (diverging)

bluescale

Blue scale color palette (sequential)

grayscale

Grayscale color palette (sequential)

Details

Pretty print and plot methods are powered by the prismatic package. Each palette is stored as a color object to enact this behaviour.

Examples

adl_palettes$primary
#> <colors>
#> #14A2FCFF 
adl_palettes$secondary
#> <colors>
#> #DBDCDDFF 
adl_palettes$tertiary
#> <colors>
#> #69DA78FF 
adl_palettes$pid3
#> <colors>
#> #14A2FCFF #60269EFF #E84C4CFF 
adl_palettes$categorical
#> <colors>
#> #14A2FCFF #DBDCDDFF #A1D55DFF #0A1A50FF #E84C4CFF #8099FFFF #FFE500FF #2C2E35FF 
adl_palettes$likert_6
#> <colors>
#> #0A1A50FF #005B98FF #14A2FCFF #FFFFFFFF #DBDCDDFF #595B60FF 
adl_palettes$likert_4
#> <colors>
#> #0A1A50FF #14A2FCFF #DBDCDDFF #595B60FF 
adl_palettes$bluescale
#> [1] "#0A1A50"   "#0A1A500D"
adl_palettes$grayscale
#> [1] "#2c2e35FF" "#2C2E350D"