Functions to access metadata

These are a set of functions that make it easy to access attributes/metadata from a vector or column in a data.frame. Note: these functions do not allow you to change, set, or remove attributes.

Usage

attr_var_label(x, df)

attr_val_labels(x, df)

attr_levels(x, df)

attr_transformation(x, df)

attr_note(x, df)

attr_question_preface(x, df)

attr_survey_flow(x, df)

Arguments

x

A vector object or the name of a column in a data.frame

df

A data.frame or tibble object. This should be specified when x is only the name of a column.

Details

attr_var_label()

This function gets the the variable label.

attr_val_labels()

This function gets the value labels.

attr_levels()

This function gets the factor levels.

attr_transformation()

This function gets the "transformation" attribute which tells you what data transformation the variable underwent when it was created.

attr_question_preface()

This function gets the "question_preface" attribute. Some questions in surveys enable respondents to select multiple responses. Each response gets it's own variable in the data, which is labeled "label. The text stored in this attribute contains the text that respondents saw before the response options.

attr_survey_flow()

This function gets teh the "survey_flow" attribute. This is used to indicate if there was an experiment or some sort of branching involved in the survey flow

attr_note()

This function gets the "note" attribute which contains miscellaneous information about the variable.

Examples

library(adlgraphs)

# get the variable labels
attr_var_label(x = test_data$edu)
#> [1] "What is the highest level of school you have completed or the highest degree you have received?"
attr_var_label(x = "edu", df =  test_data)
#> [1] "What is the highest level of school you have completed or the highest degree you have received?"

# get the value labels
attr_val_labels(x = test_data$edu)
#> Less than High School    High School or GED          Some College 
#>                     1                     2                     3 
#>     Bachelor's Degree       Graduate Degree 
#>                     4                     5 
attr_val_labels(x = "edu", df = test_data)
#> Less than High School    High School or GED          Some College 
#>                     1                     2                     3 
#>     Bachelor's Degree       Graduate Degree 
#>                     4                     5 

# get the levels for edu_f
attr_levels(x = test_data$edu_f)
#> [1] "High School or Less" "Some College"        "Bachelor's Degree"  
#> [4] "Graduate Degree"    
attr_levels(x = "edu_f", df = test_data)
#> [1] "High School or Less" "Some College"        "Bachelor's Degree"  
#> [4] "Graduate Degree"    

# get the transformation
attr_transformation(x = test_data$edu_f)
#> Recoded 'edu' as a factor and set the levels based on their order.
#> The data transformation is as follows:
#> What was 'c(1:2)' has become 'High School or Less'
#> What was '3' has become 'Some College'
#> What was '4' has become 'Bachelor's Degree'
#> What was '5' has become 'Graduate Degree'
attr_transformation(x = "edu_f", df = test_data)
#> Recoded 'edu' as a factor and set the levels based on their order.
#> The data transformation is as follows:
#> What was 'c(1:2)' has become 'High School or Less'
#> What was '3' has become 'Some College'
#> What was '4' has become 'Bachelor's Degree'
#> What was '5' has become 'Graduate Degree'