Run Dunnett's multiple comparisons test with one control.

dunnett() calculates Dunnett's post hoc pairwise multiple comparisons procedure. More simply, it calculates the mean of a variable (x) along the different levels of a grouping variable (treats) and then determines if the difference between the control/reference group and each level is statistically significant.

Usage

dunnett(
  data,
  x,
  treats = NULL,
  group = NULL,
  wt = NULL,
  control = NULL,
  conf.level = 0.95,
  show_means = FALSE
)

Arguments

data

A data frame or tibble.

x

A numeric vector that will be used to calculate the means.

treats

A variable whose values are used to determine if the means are statistically significantly different from each other.

group

A character or factor vector. It is a second level grouping variable. If specified, the function checks to see if the values in treats are statistically significant within each level of this variable.

wt

Weights. Add if you have a weighting variable and want to perform Dunnett's test with weighted means,

control

A string that specifies the level of the reference group through which the others will be tested.

conf.level

A number between 0 and 1 that signifies the width of the desired confidence interval. Default is 0.95, which corresponds to a 95% confidence interval.

show_means

Logical. Determines if the output should contain the mean differences between the levels and the reference level or if it should show the means of each level. Default is FALSE

Details

While there are other functions that also perform Dunnett's Test, like PMCMRplus::dunnettTest() and DescTools::DunnettTest() to name a few, there are a few key differences between this function and those. Firstly, this function takes in a data frame or tibble. This was done so that it can either be piped in or specified in the argument. See more in the example section below.

Another important difference, is the addition of the group variable. When group is specified, the function still performs the Dunnett test between the variable in x and the variable in treats, but it does so along each level of whatever variable is specified in group. You can see this in action in the examples section below.

A third key difference is the output. This function outputs a tibble object with a special class that enables it to be used with prettytable(). Furthermore, there are two distinct outputs. The default output shows the difference in means between each level and the control level and the confidence intervals are based on this difference. However, if show_means = TRUE, then the output shows the mean values for each level in treats while still indicating the p-value.

Lastly, this function allows you to add weights to calculate weighted means.

Examples

# Check to see if any of the education groups are significantly different
# from the control group (in this case "High School or Less") for conspiracy
# theory belief
dunnett(test_data, "acts_avg", "edu_f")
#> # A tibble: 3 × 7
#>   edu_f               diff     n conf.low conf.high p.value stars
#>   <chr>              <dbl> <dbl>    <dbl>     <dbl>   <dbl> <chr>
#> 1 Some College      -0.234    79   -0.528     0.06    0.154 ""   
#> 2 Bachelor's Degree -0.547    57   -0.868    -0.227   0     "***"
#> 3 Graduate Degree   -0.609    38   -0.972    -0.245   0     "***"
# now let's do the same but have it show the means
dunnett(test_data, "acts_avg", "edu_f", show_means = TRUE)
#> Joining with `by = join_by(edu_f)`
#> # A tibble: 4 × 8
#>   edu_f                mean    sd     n conf.low conf.high p.value stars
#>   <chr>               <dbl> <dbl> <dbl>    <dbl>     <dbl>   <dbl> <chr>
#> 1 High School or Less  2.99  0.72    76     2.82      3.16  NA      NA  
#> 2 Some College         2.76  0.72    79     2.6       2.92   0.154 ""   
#> 3 Bachelor's Degree    2.44  0.88    57     2.21      2.68   0     "***"
#> 4 Graduate Degree      2.38  0.79    38     2.12      2.64   0     "***"

# now do the same as above but make "Graduate Degree" the control group
dunnett(test_data, "acts_avg", "edu_f", control = "Graduate Degree")
#> # A tibble: 3 × 7
#>   edu_f                diff     n conf.low conf.high p.value stars
#>   <chr>               <dbl> <dbl>    <dbl>     <dbl>   <dbl> <chr>
#> 1 High School or Less 0.609    76    0.252     0.965   0     "***"
#> 2 Some College        0.375    79    0.021     0.729   0.035 "*"  
#> 3 Bachelor's Degree   0.061    57   -0.314     0.437   0.954 ""   
dunnett(test_data, "acts_avg", "edu_f", control = "Graduate Degree", show_means = TRUE)
#> Joining with `by = join_by(edu_f)`
#> # A tibble: 4 × 8
#>   edu_f                mean    sd     n conf.low conf.high p.value stars
#>   <chr>               <dbl> <dbl> <dbl>    <dbl>     <dbl>   <dbl> <chr>
#> 1 High School or Less  2.99  0.72    76     2.82      3.16   0     "***"
#> 2 Some College         2.76  0.72    79     2.6       2.92   0.035 "*"  
#> 3 Bachelor's Degree    2.44  0.88    57     2.21      2.68   0.954 ""   
#> 4 Graduate Degree      2.38  0.79    38     2.12      2.64  NA      NA  

# now let's add in partisanship (`pid_f3`) as the `group` variable. This let's us
# compare education levels within each level of `pid_f3`.
dunnett(test_data, "acts_avg", "edu_f", "pid_f3")
#> # A tibble: 9 × 8
#>   pid_f3          edu_f              diff     n conf.low conf.high p.value stars
#>   <dbl+lbl>       <chr>             <dbl> <dbl>    <dbl>     <dbl>   <dbl> <chr>
#> 1 1 [Democrat]    Some College     -0.276    32   -0.766     0.215   0.396 ""   
#> 2 1 [Democrat]    Bachelor's Degr… -0.385    28   -0.891     0.121   0.174 ""   
#> 3 1 [Democrat]    Graduate Degree  -0.577    26   -1.09     -0.062   0.024 "*"  
#> 4 2 [Independent] Some College     -0.203    22   -0.652     0.246   0.599 ""   
#> 5 2 [Independent] Bachelor's Degr… -0.426     6   -1.13      0.276   0.355 ""   
#> 6 2 [Independent] Graduate Degree  -0.218     3   -1.16      0.728   0.916 ""   
#> 7 3 [Republican]  Some College     -0.12     25   -0.669     0.429   0.921 ""   
#> 8 3 [Republican]  Bachelor's Degr… -0.613    23   -1.17     -0.053   0.028 "*"  
#> 9 3 [Republican]  Graduate Degree  -0.292     9   -1.05      0.462   0.691 ""   
# and repeat that but showing the means
dunnett(test_data, "acts_avg", "edu_f", "pid_f3", show_means = TRUE)
#> Joining with `by = join_by(pid_f3, edu_f)`
#> # A tibble: 12 × 9
#> # Groups:   pid_f3 [3]
#>    pid_f3          edu_f       mean    sd     n conf.low conf.high p.value stars
#>    <dbl+lbl>       <chr>      <dbl> <dbl> <dbl>    <dbl>     <dbl>   <dbl> <chr>
#>  1 1 [Democrat]    High Scho…  2.76  0.82    25     2.42      3.1   NA      NA  
#>  2 1 [Democrat]    Some Coll…  2.48  0.73    32     2.22      2.75   0.396 ""   
#>  3 1 [Democrat]    Bachelor'…  2.38  0.78    28     2.07      2.68   0.175 ""   
#>  4 1 [Democrat]    Graduate …  2.18  0.78    26     1.87      2.5    0.024 "*"  
#>  5 2 [Independent] High Scho…  3.13  0.53    26     2.92      3.35  NA      NA  
#>  6 2 [Independent] Some Coll…  2.93  0.54    22     2.69      3.17   0.599 ""   
#>  7 2 [Independent] Bachelor'…  2.71  0.98     6     1.68      3.74   0.355 ""   
#>  8 2 [Independent] Graduate …  2.92  1.28     3    -0.27      6.1    0.916 ""   
#>  9 3 [Republican]  High Scho…  3.07  0.77    25     2.75      3.39  NA      NA  
#> 10 3 [Republican]  Some Coll…  2.95  0.75    25     2.64      3.26   0.921 ""   
#> 11 3 [Republican]  Bachelor'…  2.46  0.98    23     2.03      2.88   0.028 "*"  
#> 12 3 [Republican]  Graduate …  2.78  0.44     9     2.44      3.12   0.691 ""