Get correlations for a combination of variables

get_all_corr makes it easy to calculate correlations across every variable in a data frame or select set of variables. It also works with grouped data frames so you can check correlations among the levels of several grouping variables.

Usage

get_all_corr(data, cols, wt = NULL, remove_redundant = TRUE)

Arguments

data

A data frame or tibble object

cols

<tidy-select> The variables you want to get the correlations for.

wt

A variable to use as the weights for weighted correlations

remove_redundant

Should rows where the two variables are the same be kept or removed? If TRUE, the default, they are removed.

Examples

# load dplyr and adlgraphs
library(dplyr)
library(adlgraphs)

# To get correlations with three variables you can do it three ways
# 1. Create a new data frame with only the columns you want
new_data <- test_data %>% dplyr::select(top:dominate)
get_all_corr(new_data)
#> # A tibble: 6 × 8
#>   x               y          correlation     n conf.low conf.high  p.value stars
#>   <chr+lbl>       <chr+lbl>        <dbl> <dbl>    <dbl>     <dbl>    <dbl> <chr>
#> 1 inferior [Some… top [An i…       0.450   250    0.339    0.562  6.89e-14 ***  
#> 2 dominate [No o… top [An i…      -0.191   250   -0.314   -0.0686 2.38e- 3 **   
#> 3 top [An ideal … inf… [Som…       0.450   250    0.339    0.562  6.89e-14 ***  
#> 4 dominate [No o… inf… [Som…      -0.195   250   -0.318   -0.0728 1.91e- 3 **   
#> 5 top [An ideal … dom… [No …      -0.191   250   -0.314   -0.0686 2.38e- 3 **   
#> 6 inferior [Some… dom… [No …      -0.195   250   -0.318   -0.0728 1.91e- 3 **   

# 2. Using dplyr::select() and pipes
test_data %>% 
  dplyr::select(c(top:dominate)) %>% 
  get_all_corr()
#> # A tibble: 6 × 8
#>   x               y          correlation     n conf.low conf.high  p.value stars
#>   <chr+lbl>       <chr+lbl>        <dbl> <dbl>    <dbl>     <dbl>    <dbl> <chr>
#> 1 inferior [Some… top [An i…       0.450   250    0.339    0.562  6.89e-14 ***  
#> 2 dominate [No o… top [An i…      -0.191   250   -0.314   -0.0686 2.38e- 3 **   
#> 3 top [An ideal … inf… [Som…       0.450   250    0.339    0.562  6.89e-14 ***  
#> 4 dominate [No o… inf… [Som…      -0.195   250   -0.318   -0.0728 1.91e- 3 **   
#> 5 top [An ideal … dom… [No …      -0.191   250   -0.314   -0.0686 2.38e- 3 **   
#> 6 inferior [Some… dom… [No …      -0.195   250   -0.318   -0.0728 1.91e- 3 **   

# 3. Use the `cols` argument
get_all_corr(test_data, cols = c(top:dominate))
#> # A tibble: 6 × 8
#>   x               y          correlation     n conf.low conf.high  p.value stars
#>   <chr+lbl>       <chr+lbl>        <dbl> <dbl>    <dbl>     <dbl>    <dbl> <chr>
#> 1 inferior [Some… top [An i…       0.450   250    0.339    0.562  6.89e-14 ***  
#> 2 dominate [No o… top [An i…      -0.191   250   -0.314   -0.0686 2.38e- 3 **   
#> 3 top [An ideal … inf… [Som…       0.450   250    0.339    0.562  6.89e-14 ***  
#> 4 dominate [No o… inf… [Som…      -0.195   250   -0.318   -0.0728 1.91e- 3 **   
#> 5 top [An ideal … dom… [No …      -0.191   250   -0.314   -0.0686 2.38e- 3 **   
#> 6 inferior [Some… dom… [No …      -0.195   250   -0.318   -0.0728 1.91e- 3 **   
# or 
test_data %>% get_all_corr(c(top:dominate))
#> # A tibble: 6 × 8
#>   x               y          correlation     n conf.low conf.high  p.value stars
#>   <chr+lbl>       <chr+lbl>        <dbl> <dbl>    <dbl>     <dbl>    <dbl> <chr>
#> 1 inferior [Some… top [An i…       0.450   250    0.339    0.562  6.89e-14 ***  
#> 2 dominate [No o… top [An i…      -0.191   250   -0.314   -0.0686 2.38e- 3 **   
#> 3 top [An ideal … inf… [Som…       0.450   250    0.339    0.562  6.89e-14 ***  
#> 4 dominate [No o… inf… [Som…      -0.195   250   -0.318   -0.0728 1.91e- 3 **   
#> 5 top [An ideal … dom… [No …      -0.191   250   -0.314   -0.0686 2.38e- 3 **   
#> 6 inferior [Some… dom… [No …      -0.195   250   -0.318   -0.0728 1.91e- 3 **   

# To get weighted correlations just specify the `wt` argument
test_data %>% get_all_corr(c(top:dominate), wt = wts)
#> # A tibble: 6 × 8
#>   x               y          correlation     n conf.low conf.high  p.value stars
#>   <chr+lbl>       <chr+lbl>        <dbl> <dbl>    <dbl>     <dbl>    <dbl> <chr>
#> 1 inferior [Some… top [An i…       0.450   250    0.339    0.562  6.89e-14 ***  
#> 2 dominate [No o… top [An i…      -0.191   250   -0.314   -0.0686 2.38e- 3 **   
#> 3 top [An ideal … inf… [Som…       0.450   250    0.339    0.562  6.89e-14 ***  
#> 4 dominate [No o… inf… [Som…      -0.195   250   -0.318   -0.0728 1.91e- 3 **   
#> 5 top [An ideal … dom… [No …      -0.191   250   -0.314   -0.0686 2.38e- 3 **   
#> 6 inferior [Some… dom… [No …      -0.195   250   -0.318   -0.0728 1.91e- 3 **   

# You can also calculate grouped correlations. For example, if
# you were interested in comparing the weighted correlations 
# among people with a college degree vs those without one, you 
# would do it like this:
test_data %>% 
  dplyr::group_by(edu_f2) %>% 
  get_all_corr(c(top:dominate), wt = wts)
#> # A tibble: 12 × 9
#>    edu_f2    x          y          correlation     n conf.low conf.high  p.value
#>    <chr>     <chr+lbl>  <chr+lbl>        <dbl> <dbl>    <dbl>     <dbl>    <dbl>
#>  1 At Least… inf… [Som… top [An i…       0.313    95    0.118    0.509  2.00e- 3
#>  2 At Least… dom… [No … top [An i…      -0.348    95   -0.541   -0.155  5.51e- 4
#>  3 At Least… top [An i… inf… [Som…       0.313    95    0.118    0.509  2.00e- 3
#>  4 At Least… dom… [No … inf… [Som…      -0.298    95   -0.494   -0.101  3.38e- 3
#>  5 At Least… top [An i… dom… [No …      -0.348    95   -0.541   -0.155  5.51e- 4
#>  6 At Least… inf… [Som… dom… [No …      -0.298    95   -0.494   -0.101  3.38e- 3
#>  7 No Colle… inf… [Som… top [An i…       0.478   155    0.337    0.618  3.27e-10
#>  8 No Colle… dom… [No … top [An i…      -0.121   155   -0.279    0.0378 1.34e- 1
#>  9 No Colle… top [An i… inf… [Som…       0.478   155    0.337    0.618  3.27e-10
#> 10 No Colle… dom… [No … inf… [Som…      -0.142   155   -0.300    0.0163 7.83e- 2
#> 11 No Colle… top [An i… dom… [No …      -0.121   155   -0.279    0.0378 1.34e- 1
#> 12 No Colle… inf… [Som… dom… [No …      -0.142   155   -0.300    0.0163 7.83e- 2
#> # ℹ 1 more variable: stars <chr>