Calculate the loadings in factor analysis

This function creates a data frame of factor loadings from a factor analysis. It can be on an object that was created with psych::fa() or it can be used on a data frame. If used on a data frame, it will run it on all columns in the data frame. Also works on grouped data frame if you want to check how the factor loadings may change along the different levels of a group.

Usage

get_loadings(
  model,
  labels = NULL,
  threshold = 0.4,
  print = "short",
  nfactors = 1,
  fm = "pa",
  rotate = "oblimin"
)

Arguments

model

Either a model created using psych::fa() or a data.frame

labels

Either a character vector or data frame. Creates a new column called "labels" for each variable in the factor analysis.

threshold

The threshold with which to not show the factor loadings. Default is 0.4.

print

The printing method. Default is "short" which only prints a dataframe of the factor loadings. The alternative is "long" but that has not been created yet.

nfactors

Number of factors to extract, default is 1.

fm

Factoring method fm="minres" will do a minimum residual as will fm="uls". Both of these use a first derivative. fm="ols" differs very slightly from "minres" in that it minimizes the entire residual matrix using an OLS procedure but uses the empirical first derivative. This will be slower. fm="wls" will do a weighted least squares (WLS) solution, fm="gls" does a generalized weighted least squares (GLS), fm="pa" will do the principal factor solution, fm="ml" will do a maximum likelihood factor analysis. fm="minchi" will minimize the sample size weighted chi square when treating pairwise correlations with different number of subjects per pair. fm ="minrank" will do a minimum rank factor analysis. "old.min" will do minimal residual the way it was done prior to April, 2017 (see discussion below). fm="alpha" will do alpha factor analysis as described in Kaiser and Coffey (1965). Default is "pa".

rotate

"none", "varimax", "quartimax", "bentlerT", "equamax", "varimin", "geominT" and "bifactor" are orthogonal rotations. "Promax", "promax", "oblimin", "simplimax", "bentlerQ, "geominQ" and "biquartimin" and "cluster" are possible oblique transformations of the solution. The default is to do a oblimin transformation, although versions prior to 2009 defaulted to varimax. SPSS seems to do a Kaiser normalization before doing Promax, this is done here by the call to "promax" which does the normalization before calling Promax in GPArotation.