Make dichotomous factors

Create a vector of a dichotomous factor. Was designed to work on any vector that is of class "factor" or "haven_labelled". However, if the vector is numeric with no value labels, the function will return an error. In addition, this function works by removing the first from the factors, unless it is only one word. To illustrate what I mean, if a factor vector contains "Strongly agree", the "Strongly" gets removed so it is just "Agree". If your vector is not set up this way, then this function will not work properly.

Usage

make_dicho(x)

Arguments

x

A vector of type "haven_labelled" or "factor"

Details

In addition, this function adds two new attributes. The first attribute, transformation, indicates the data transformation that the original vector underwent to create this new vector. The second attribute, label, contains the variable label that was found in the original variable. However, if the original vector did not have a variable label, then this attribute will not show up.

Examples

library(tibble)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(labelled)
library(haven)

# create fake data
data <- tibble::tribble(
  ~x, ~y, ~z,
  3, 2, 3,
  4, 4, 2,
  2, 6, 1,
  1, 1, 4,
  5, 4, 3,
  6, 5, 6
  ) %>%
  # add value labels
  labelled::set_value_labels(
    x = c(`Strongly agree` = 1,
          `Agree` = 2,
          `Somewhat agree` = 3,
          `Somewhat disagree` = 4,
          `Disagree` = 5,
          `Strongly disagree` = 6),
    y = c(`Strongly agree` = 1,
          `Agree` = 2,
          `Somewhat agree` = 3,
          `Somewhat disagree` = 4,
          `Disagree` = 5,
          `Strongly disagree` = 6),
    z = c(`Strongly agree` = 1,
          `Agree` = 2,
          `Somewhat agree` = 3,
          `Somewhat disagree` = 4,
          `Disagree` = 5,
          `Strongly disagree` = 6)
  )

 # show the data transformation with a haven_labelled vector
 new_data <- data %>% dplyr::mutate(new_x = make_dicho(x))
#> Error in dplyr::mutate(., new_x = make_dicho(x)): ℹ In argument: `new_x = make_dicho(x)`.
#> Caused by error in `make_dicho()`:
#> ! could not find function "make_dicho"
 # check the updated dataset
 new_data
#> Error in eval(expr, envir, enclos): object 'new_data' not found

 # Check the attributes
 attributes(new_data$new_x)
#> Error in eval(expr, envir, enclos): object 'new_data' not found
 # another way of checking the attributes
 str(new_data$new_x)
#> Error in eval(expr, envir, enclos): object 'new_data' not found

 # check the factor levels
 unique(new_data$new_x)
#> Error in eval(expr, envir, enclos): object 'new_data' not found

 ---------------------------------------------------------------------------
 # when done on a factor vector, the function still works
 new_data <- data %>%
   dplyr::mutate(
     # convert x to a factor and name it "factor_x"
     factor_x = haven::as_factor(x),
     # convert the x_factor variable to a dichotomous factor
     new_x = make_dicho(x_factor)
   )
#> Error in dplyr::mutate(., factor_x = haven::as_factor(x), new_x = make_dicho(x_factor)): ℹ In argument: `new_x = make_dicho(x_factor)`.
#> Caused by error in `make_dicho()`:
#> ! could not find function "make_dicho"
 # check the updated dataset
 new_data
#> Error in eval(expr, envir, enclos): object 'new_data' not found

 # Check the attributes
 attributes(new_data$new_x)
#> Error in eval(expr, envir, enclos): object 'new_data' not found
 # another way of checking the attributes
 str(new_data$new_x)
#> Error in eval(expr, envir, enclos): object 'new_data' not found

 # check the factor levels
 unique(new_data$new_x)
#> Error in eval(expr, envir, enclos): object 'new_data' not found

----------------------------------------------------------------------------
# function also works inside dplyr::across()
 new_data <- data %>%
   dplyr::mutate(
     dplyr::across(x:z, make_dicho, .names = "new{col}")
   )
#> Error in dplyr::mutate(., dplyr::across(x:z, make_dicho, .names = "new{col}")): ℹ In argument: `dplyr::across(x:z, make_dicho, .names = "new{col}")`.
#> Caused by error:
#> ! object 'make_dicho' not found
 # check the updated dataset
 new_data
#> Error in eval(expr, envir, enclos): object 'new_data' not found

 # check the attributes for the new dataset
 str(new_data)
#> Error in eval(expr, envir, enclos): object 'new_data' not found