Reverse a numeric function

Reverse a numeric variable while maintaining variable and value labels if available. Also adds an attribute describing the transformation the original variable underwent.

Usage

num_rev(x)

Arguments

x

A vector of class haven_labelled or numeric

Examples

library(dplyr)
library(labelled)
library(adlgraphs)
# create fake data
df <- tibble::tibble(
  w = sample(1:4, 10, replace = TRUE),
  x = sample(1:4, 10, replace = TRUE),
  y = sample(1:4, 10, replace = TRUE),
  z = sample(1:4, 10, replace = TRUE)
) %>%
  # add value labels to x and y but not z to see what happens with no labels
  labelled::set_value_labels(
    x = c(`Strongly agree` = 1,
          `Somewhat agree` = 2,
          `Somewhat disagree` = 3,
          `Strongly disagree` = 4),
    y = c(`Strongly agree` = 1,
          `Somewhat agree` = 2,
          `Somewhat disagree` = 3,
          `Strongly disagree` = 4),
  ) %>%
  # set variable labels to x and z
  labelled::set_variable_labels(
    x = "This is the variable label for x",
    z = "This is the variable label for z"
  )

rev_df <- df %>% mutate(rev_w = num_rev(w))

head(rev_df)
#> # A tibble: 6 × 5
#>       w x                     y                         z rev_w
#>   <int> <int+lbl>             <int+lbl>             <int> <dbl>
#> 1     1 2 [Somewhat agree]    3 [Somewhat disagree]     3     4
#> 2     3 1 [Strongly agree]    3 [Somewhat disagree]     3     2
#> 3     4 1 [Strongly agree]    2 [Somewhat agree]        4     1
#> 4     3 1 [Strongly agree]    3 [Somewhat disagree]     4     2
#> 5     1 3 [Somewhat disagree] 1 [Strongly agree]        2     4
#> 6     2 4 [Strongly disagree] 1 [Strongly agree]        3     3

str(rev_df)
#> tibble [10 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ w    : int [1:10] 1 3 4 3 1 2 3 3 3 4
#>  $ x    : int+lbl [1:10] 2, 1, 1, 1, 3, 4, 3, 1, 2, 2
#>    ..@ labels: Named int [1:4] 1 2 3 4
#>    .. ..- attr(*, "names")= chr [1:4] "Strongly agree" "Somewhat agree" "Somewhat disagree" "Strongly disagree"
#>    ..@ label : chr "This is the variable label for x"
#>  $ y    : int+lbl [1:10] 3, 3, 2, 3, 1, 1, 4, 1, 2, 2
#>    ..@ labels: Named int [1:4] 1 2 3 4
#>    .. ..- attr(*, "names")= chr [1:4] "Strongly agree" "Somewhat agree" "Somewhat disagree" "Strongly disagree"
#>  $ z    : int [1:10] 3 3 4 4 2 3 1 2 3 3
#>   ..- attr(*, "label")= chr "This is the variable label for z"
#>  $ rev_w: num [1:10] 4 2 1 2 4 3 2 2 2 1
#>   ..- attr(*, "transformation")= 'glue' chr "Reversing 'w'"