Stat for density ridgeline plots

This stat is the default stat used by geom_density_quant. It is very similar to stat_density and stat_density_ridges as it was built as a sort of combination of the two. One of the key differences between this function and those two is that this one uses the Sheather & Jones ("sj") as the default bandwidth selector. This is done because this is a better bandwidth selector than Silverman's ("nrd0") which is the default for the other two functions. In addition, this function allows you to add quantile lines similar to stat_density_ridges.

Usage

stat_density_quant(
  mapping = NULL,
  data = NULL,
  geom = geom,
  position = "stack",
  ...,
  bw = "sj",
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  bounds = c(-Inf, Inf),
  show.legend = NA,
  inherit.aes = TRUE,
  quantile_lines = FALSE,
  calc_ecdf = FALSE,
  quantiles = 4,
  quantile_fun = quantile
)

Arguments

geom

The geometric object to use to display the data. geom_density_quant is the default.

bw

The smoothing bandwidth to be used. If numeric, the standard deviation of the smoothing kernel. If character, a rule to choose the bandwidth, as listed in stats::bw.nrd(). Note that automatic calculation of the bandwidth does not take weights into account. Default is sj.

adjust

A multiplicate bandwidth adjustment. This makes it possible to adjust the bandwidth while still using the a bandwidth estimator. For example, adjust = 1/2 means use half of the default bandwidth.

kernel

Kernel. See list of available kernels in density().

n

The number of equally spaced points at which the density is to be estimated. Should be a power of 2. Default is 512.

bounds

Known lower and upper bounds for estimated data. Default c(-Inf, Inf) means that there are no (finite) bounds. If any bound is finite, boundary effect of default density estimation will be corrected by reflecting tails outside bounds around their closest edge. Data points outside of bounds are removed with a warning.

calc_ecdf

If TRUE, stat_density_ridges calculates an empirical cumulative distribution function (ecdf) and returns a variable ecdf and a variable quantile. Both can be mapped onto aesthetics via stat(ecdf) and stat(quantile), respectively.

quantiles

Sets the number of quantiles the data should be broken into. Used if either calc_ecdf = TRUE or quantile_lines = TRUE. If quantiles is an integer then the data will be cut into that many equal quantiles. If it is a vector of probabilities then the data will cut by them.

quantile_fun

Function that calculates quantiles. The function needs to accept two parameters, a vector x holding the raw data values and a vector probs providing the probabilities that define the quantiles. Default is quantile.

from, to

The left and right-most points of the grid at which the density is to be estimated, as in density(). If not provided, these are estimated from the data range and the bandwidth. and sets it to TRUE.