dataset_title,

dataset_description,

Usage

test_data

Format

A data frame with nrow(df) rows and ncol(df) variables:

resp_id

Unique ID for each Respondent

wts

Survey weighting variable

edu

What is the highest level of school you have completed or the highest degree you have received?

pid_f3

Political Partisanship

pol_part_rally

Attended a political rally, speech or campaign event

pol_part_worked

Worked or volunteered for a political party, candidate or campaign

pol_part_contact

Contacted an elected official

pol_part_money

Contributed money to a candidate running for public office or to a group working to elect a candidate

pol_part_social

Publicly expressed your support for a political candidate on Facebook, Twitter/X, Instagram, or other social media

pol_part_attended

Attended government meetings in your community, such as city or town council meetings

pol_part_none

None of the above

top

An ideal society requires some groups to be on top and others to be on the bottom

inferior

Some groups of people are simply inferior to other groups

dominate

No one group should dominate in society

deserving

Groups at the bottom are just as deserving as groups at the top

special

It is unfair for some groups in society to receive special treatment from the government

harder

I have a harder time succeeding than my parents did

controlled

Much of our lives are being controlled by plots hatched in secrecy

small

Even though we live in a democracy, a few people will always run things anyway

run

The people who really "run" the country are not known to the voters

big_events

Big events like wars, recessions, and the outcomes of elections are controlled by small groups of people who are working in secret against the rest of us

accept_hamas

If you had a close family member who supported Hamas, would your family and friends find it…

accept_isr

If you had a close family member who supported Israel, would your family and friends find it…

dislike_jews

How many of your close friends/family dislike Jews

stick_together

Jews stick together more than other Americans

values

Jews share my values.

head

Jews always like to be at the head of things.

loyal

Jews are more loyal to Israel than to America.

business_power

Jews have too much power in the business world.

wall_street

Jews have too much control and influence on Wall Street.

trad_n

ADL Index

edu_f

What is the highest level of school you have completed or the highest degree you have received?

edu_f2

What is the highest level of school you have completed or the highest degree you have received?

top_rev

An ideal society requires some groups to be on top and others to be on the bottom

inferior_rev

Some groups of people are simply inferior to other groups

dominate_flip

One group should dominate in society

deserving_flip

Groups at the bottom are not just as deserving as groups at the top

sdo_sum

Social Dominance Orientation

sdo_avg

Social Dominance Orientation

pol_part_sum

Political Participation

controlled_rev

Much of our lives are being controlled by plots hatched in secrecy

small_rev

Even though we live in a democracy, a few people will always run things anyway

run_rev

The people who really "run" the country are not known to the voters

big_events_rev

Big events like wars, recessions, and the outcomes of elections are controlled by small groups of people who are working in secret against the rest of us

acts_sum

Conspiracy Theory Belief

acts_avg

Conspiracy Theory Belief