Calculate weighted frequencies

Use this function to calculate simple weighted frequencies weighted grouped. You can also specify a grouping variable by which you want to calculate the frequencies

Usage

get_freqs(df, x, group, wt, cross_tab = FALSE)

Arguments

df

An object of type data.frame or tibble. If piping the df into the function, this is not required.

x

Either a character string or symbol. The variable with which want to get the frequencies.

group

Either a character string or a symbol. The grouping variable.

wt

Weights. Add if you have a weighting variable and want to get weighted frequencies

cross_tab

Logical. If a group object has been supplied, should the the table be pivoted to create make it like crosstabs

Details

The x, group, and wt arguments can either be strings or symbols (meaning they can have quotes or no quotes). The benefit of this is that it makes it really easy to iterate this function over a list or vector of variables with other functions like map() purrr::map() or walk() purrr::walk() that are found in the purrr package.

Examples

# load the package
library(dplyr)

# Let's calculate the overall frequency for trad_n
get_freqs(test_data, trad_n)
#> # A tibble: 15 × 3
#>    trad_n     n    pct
#>     <dbl> <int>  <dbl>
#>  1      0    31 0.150 
#>  2      1    23 0.111 
#>  3      2    27 0.130 
#>  4      3    34 0.164 
#>  5      4    15 0.0725
#>  6      5    14 0.0676
#>  7      6     7 0.0338
#>  8      7     7 0.0338
#>  9      8     6 0.0290
#> 10      9     3 0.0145
#> 11     10     6 0.0290
#> 12     11     8 0.0386
#> 13     12     4 0.0193
#> 14     13    12 0.0580
#> 15     14    10 0.0483

# it also works if x is a string
get_means(test_data, "trad_n")
#> # A tibble: 1 × 6
#>    mean    sd     n std.error conf.low conf.high
#>   <dbl> <dbl> <dbl>     <dbl>    <dbl>     <dbl>
#> 1  4.67  4.33   207       0.3     4.08      5.26

# Let's do that again but add weights
get_freqs(test_data, trad_n, wt = wts)
#> # A tibble: 15 × 3
#>    trad_n     n    pct
#>     <dbl> <dbl>  <dbl>
#>  1      0  27.6 0.134 
#>  2      1  23.8 0.116 
#>  3      2  27.4 0.133 
#>  4      3  34.1 0.166 
#>  5      4  15.1 0.0730
#>  6      5  15.8 0.0767
#>  7      6   9.9 0.0481
#>  8      7   6.8 0.0329
#>  9      8   3.1 0.0152
#> 10      9   6.1 0.0296
#> 11     10   5.8 0.0282
#> 12     11   9.5 0.0462
#> 13     12   2.5 0.0119
#> 14     13  10.2 0.0492
#> 15     14   8.4 0.0407

# the wt argument can also be in quotes like this
get_freqs(test_data, "trad_n", wt = "wts")
#> # A tibble: 15 × 3
#>    trad_n     n    pct
#>     <dbl> <dbl>  <dbl>
#>  1      0  27.6 0.134 
#>  2      1  23.8 0.116 
#>  3      2  27.4 0.133 
#>  4      3  34.1 0.166 
#>  5      4  15.1 0.0730
#>  6      5  15.8 0.0767
#>  7      6   9.9 0.0481
#>  8      7   6.8 0.0329
#>  9      8   3.1 0.0152
#> 10      9   6.1 0.0296
#> 11     10   5.8 0.0282
#> 12     11   9.5 0.0462
#> 13     12   2.5 0.0119
#> 14     13  10.2 0.0492
#> 15     14   8.4 0.0407

# Now let's do the average score for different education levels
get_freqs(test_data, trad_n, edu_f, wts)
#> # A tibble: 48 × 4
#> # Groups:   edu_f [4]
#>    edu_f               trad_n     n    pct
#>    <fct>                <dbl> <dbl>  <dbl>
#>  1 High School or Less      0   5.9 0.0781
#>  2 High School or Less      1   5.3 0.0695
#>  3 High School or Less      2  12.1 0.159 
#>  4 High School or Less      3  11.5 0.152 
#>  5 High School or Less      4   4   0.0532
#>  6 High School or Less      5   6.4 0.0850
#>  7 High School or Less      6   7.6 0.101 
#>  8 High School or Less      7   1.4 0.0189
#>  9 High School or Less      8   1.6 0.0209
#> 10 High School or Less      9   5.1 0.0672
#> # ℹ 38 more rows

# it also works with quotes
get_freqs(test_data, "trad_n", "edu_f", "wts")
#> # A tibble: 48 × 4
#> # Groups:   edu_f [4]
#>    edu_f               trad_n     n    pct
#>    <fct>                <dbl> <dbl>  <dbl>
#>  1 High School or Less      0   5.9 0.0781
#>  2 High School or Less      1   5.3 0.0695
#>  3 High School or Less      2  12.1 0.159 
#>  4 High School or Less      3  11.5 0.152 
#>  5 High School or Less      4   4   0.0532
#>  6 High School or Less      5   6.4 0.0850
#>  7 High School or Less      6   7.6 0.101 
#>  8 High School or Less      7   1.4 0.0189
#>  9 High School or Less      8   1.6 0.0209
#> 10 High School or Less      9   5.1 0.0672
#> # ℹ 38 more rows

# if we want to pivot the results so they look like cross tabs, then we need
# to set `cross_tab` to TRUE
get_freqs(test_data, trad_n, edu_f, wts, cross_tab = TRUE)
#> # A tibble: 15 × 5
#>    trad_n `High School or Less` `Some College`    `Bachelor's Degree`
#>     <dbl> <glue>                <glue>            <glue>             
#>  1      0 7.81% (n = 5.9)       18.92% (n = 14.4) 14.6% (n = 4.9)    
#>  2      1 6.95% (n = 5.3)       15.46% (n = 11.7) 10.48% (n = 3.5)   
#>  3      2 15.93% (n = 12.1)     7% (n = 5.3)      15.83% (n = 5.3)   
#>  4      3 15.22% (n = 11.5)     17.79% (n = 13.5) 12.24% (n = 4.1)   
#>  5      4 5.32% (n = 4)         8.41% (n = 6.4)   10.74% (n = 3.6)   
#>  6      5 8.5% (n = 6.4)        4.84% (n = 3.7)   13.02% (n = 4.4)   
#>  7      6 10.07% (n = 7.6)      1.94% (n = 1.5)   2.43% (n = 0.8)    
#>  8      7 1.89% (n = 1.4)       4.69% (n = 3.6)   5.39% (n = 1.8)    
#>  9      8 2.09% (n = 1.6)       1.54% (n = 1.2)   1.13% (n = 0.4)    
#> 10      9 6.72% (n = 5.1)       NA                NA                 
#> 11     10 1.95% (n = 1.5)       4.33% (n = 3.3)   NA                 
#> 12     11 6.13% (n = 4.6)       6.42% (n = 4.9)   NA                 
#> 13     12 1.86% (n = 1.4)       NA                NA                 
#> 14     13 5.25% (n = 4)         3.82% (n = 2.9)   9.79% (n = 3.3)    
#> 15     14 4.32% (n = 3.3)       4.82% (n = 3.7)   4.36% (n = 1.5)    
#> # ℹ 1 more variable: `Graduate Degree` <glue>

# you can also pipe in the `df` argument if you want to do some data
# transformations before you calculate the means. For example, say you want
# to compare the means of `trad_n` among people who agreed vs disagreed with
# the variable `prod_isr`:
test_data %>%
  mutate(prod_isr_f2 = make_dicho(prod_isr)) %>%
  get_freqs(trad_n, prod_isr_f2, wts)
#> # A tibble: 30 × 4
#> # Groups:   prod_isr_f2 [2]
#>    prod_isr_f2 trad_n     n     pct
#>    <fct>        <dbl> <dbl>   <dbl>
#>  1 Agree            0  23.2 0.159  
#>  2 Agree            1  21.8 0.149  
#>  3 Agree            2  18.9 0.129  
#>  4 Agree            3  27.5 0.188  
#>  5 Agree            4  13.6 0.0931 
#>  6 Agree            5  11.2 0.0770 
#>  7 Agree            6   3.6 0.0247 
#>  8 Agree            7   4.3 0.0295 
#>  9 Agree            8   1.2 0.00824
#> 10 Agree            9   2.8 0.0192 
#> # ℹ 20 more rows