Calculate weighted frequencies

Use this function to calculate simple weighted frequencies weighted grouped. You can also specify a grouping variable by which you want to calculate the frequencies

Usage

get_freqs(df, x, group, wt, cross_tab = FALSE)

Arguments

df

An object of type data.frame or tibble. If piping the df into the function, this is not required.

x

Either a character string or symbol. The variable with which want to get the frequencies.

group

Either a character string or a symbol. The grouping variable.

wt

Weights. Add if you have a weighting variable and want to get weighted frequencies

cross_tab

Logical. If a group object has been supplied, should the the table be pivoted to create make it like crosstabs

Details

The x, group, and wt arguments can either be strings or symbols (meaning they can have quotes or no quotes). The benefit of this is that it makes it really easy to iterate this function over a list or vector of variables with other functions like map() purrr::map() or walk() purrr::walk() that are found in the purrr package.

Examples

# load the package
library(dplyr)

# Let's calculate the overall frequency for big_events
get_freqs(test_data, big_events)
#> # A tibble: 4 × 3
#>   big_events                n   pct
#> * <dbl+lbl>             <int> <dbl>
#> 1 1 [Strongly agree]       39 0.156
#> 2 2 [Somewhat agree]       89 0.356
#> 3 3 [Somewhat disagree]    66 0.264
#> 4 4 [Strongly disagree]    56 0.224

# it also works if x is a string
get_means(test_data, "big_events")
#> # A tibble: 1 × 6
#>    mean    sd     n std.error conf.low conf.high
#>   <dbl> <dbl> <dbl>     <dbl>    <dbl>     <dbl>
#> 1  2.56  1.01   250      0.06     2.43      2.68

# Let's do that again but add weights
get_freqs(test_data, big_events, wt = wts)
#> # A tibble: 4 × 3
#>   big_events                n   pct
#> * <dbl+lbl>             <dbl> <dbl>
#> 1 1 [Strongly agree]     37.8 0.144
#> 2 2 [Somewhat agree]     87.6 0.333
#> 3 3 [Somewhat disagree]  73.7 0.280
#> 4 4 [Strongly disagree]  63.8 0.243

# the wt argument can also be in quotes like this
get_freqs(test_data, "big_events", wt = "wts")
#> # A tibble: 4 × 3
#>   big_events                n   pct
#> * <dbl+lbl>             <dbl> <dbl>
#> 1 1 [Strongly agree]     37.8 0.144
#> 2 2 [Somewhat agree]     87.6 0.333
#> 3 3 [Somewhat disagree]  73.7 0.280
#> 4 4 [Strongly disagree]  63.8 0.243

# Now let's do the average score for different education levels
get_freqs(test_data, big_events, edu_f, wts)
#> # A tibble: 16 × 4
#> # Groups:   edu_f [4]
#>    edu_f               big_events                n    pct
#>  * <fct>               <dbl+lbl>             <dbl>  <dbl>
#>  1 High School or Less 1 [Strongly agree]      6.2 0.0733
#>  2 High School or Less 2 [Somewhat agree]     33.6 0.398 
#>  3 High School or Less 3 [Somewhat disagree]  28.7 0.339 
#>  4 High School or Less 4 [Strongly disagree]  16.1 0.190 
#>  5 Some College        1 [Strongly agree]     21.1 0.197 
#>  6 Some College        2 [Somewhat agree]     33.3 0.311 
#>  7 Some College        3 [Somewhat disagree]  22   0.206 
#>  8 Some College        4 [Strongly disagree]  30.5 0.286 
#>  9 Bachelor's Degree   1 [Strongly agree]      7.1 0.185 
#> 10 Bachelor's Degree   2 [Somewhat agree]     12.3 0.321 
#> 11 Bachelor's Degree   3 [Somewhat disagree]  12.2 0.319 
#> 12 Bachelor's Degree   4 [Strongly disagree]   6.7 0.175 
#> 13 Graduate Degree     1 [Strongly agree]      3.5 0.105 
#> 14 Graduate Degree     2 [Somewhat agree]      8.5 0.254 
#> 15 Graduate Degree     3 [Somewhat disagree]  10.8 0.325 
#> 16 Graduate Degree     4 [Strongly disagree]  10.5 0.316 

# it also works with quotes
get_freqs(test_data, "big_events", "edu_f", "wts")
#> # A tibble: 16 × 4
#> # Groups:   edu_f [4]
#>    edu_f               big_events                n    pct
#>  * <fct>               <dbl+lbl>             <dbl>  <dbl>
#>  1 High School or Less 1 [Strongly agree]      6.2 0.0733
#>  2 High School or Less 2 [Somewhat agree]     33.6 0.398 
#>  3 High School or Less 3 [Somewhat disagree]  28.7 0.339 
#>  4 High School or Less 4 [Strongly disagree]  16.1 0.190 
#>  5 Some College        1 [Strongly agree]     21.1 0.197 
#>  6 Some College        2 [Somewhat agree]     33.3 0.311 
#>  7 Some College        3 [Somewhat disagree]  22   0.206 
#>  8 Some College        4 [Strongly disagree]  30.5 0.286 
#>  9 Bachelor's Degree   1 [Strongly agree]      7.1 0.185 
#> 10 Bachelor's Degree   2 [Somewhat agree]     12.3 0.321 
#> 11 Bachelor's Degree   3 [Somewhat disagree]  12.2 0.319 
#> 12 Bachelor's Degree   4 [Strongly disagree]   6.7 0.175 
#> 13 Graduate Degree     1 [Strongly agree]      3.5 0.105 
#> 14 Graduate Degree     2 [Somewhat agree]      8.5 0.254 
#> 15 Graduate Degree     3 [Somewhat disagree]  10.8 0.325 
#> 16 Graduate Degree     4 [Strongly disagree]  10.5 0.316 

# if we want to pivot the results so they look like cross tabs, then we need
# to set `cross_tab` to TRUE
get_freqs(test_data, big_events, edu_f, wts, cross_tab = TRUE)
#> # A tibble: 4 × 5
#>   big_events            `High School or Less` `Some College` `Bachelor's Degree`
#> * <dbl+lbl>             <glue>                <glue>         <glue>             
#> 1 1 [Strongly agree]    7.33% (n = 6.2)       19.7% (n = 21… 18.5% (n = 7.1)    
#> 2 2 [Somewhat agree]    39.77% (n = 33.6)     31.11% (n = 3… 32.12% (n = 12.3)  
#> 3 3 [Somewhat disagree] 33.88% (n = 28.7)     20.62% (n = 2… 31.86% (n = 12.2)  
#> 4 4 [Strongly disagree] 19.02% (n = 16.1)     28.57% (n = 3… 17.53% (n = 6.7)   
#> # ℹ 1 more variable: `Graduate Degree` <glue>

# you can also pipe in the `df` argument if you want to do some data
# transformations before you calculate the means. For example, say you want
# to compare the frequencies of `big_events` among people who agreed vs
# disagreed with the variable `top`:
test_data %>%
  mutate(top_f2 = make_dicho(top)) %>%
  get_freqs(trad_n, top_f2, wts)
#> # A tibble: 29 × 4
#> # Groups:   top_f2 [2]
#>    top_f2 trad_n     n    pct
#>  * <fct>   <dbl> <dbl>  <dbl>
#>  1 Agree       0   4.9 0.0512
#>  2 Agree       1  10.4 0.107 
#>  3 Agree       2  13.9 0.144 
#>  4 Agree       3  11.2 0.116 
#>  5 Agree       4  11.4 0.118 
#>  6 Agree       5   2.4 0.0253
#>  7 Agree       6   8.4 0.0871
#>  8 Agree       7   2.6 0.0267
#>  9 Agree       8   3.1 0.0326
#> 10 Agree       9   5.9 0.0614
#> # ℹ 19 more rows