Calculate means with confidence intervals

Use this function to calculate simple weighted means with 95% confidence intervals or weighted grouped means.

Usage

get_means(df, x, group, wt)

Arguments

df

An object of type data.frame or tibble. If piping the df into the function, this is not required.

x

Either a character string or symbol. The variable with which you want to get the mean.

group

Either a character string or a symbol. The variable you want the means to be grouped by.

wt

Weights. Add if you have a weighting variable and want to get weighted means

Details

The x, group, and wt arguments can either be strings or symbols (meaning they can have quotes or no quotes). The benefit of this is that it makes it really easy to iterate this function over a list or vector of variables with other functions like map() purrr::map() or walk() purrr::walk() that are found in the purrr package.

Examples

# load the package
library(dplyr)

# Let's calculate the overall average score for trad_n
get_means(test_data, trad_n)
#> # A tibble: 1 × 6
#>    mean    sd     n std.error conf.low conf.high
#>   <dbl> <dbl> <dbl>     <dbl>    <dbl>     <dbl>
#> 1  4.38  3.66   250      0.23     3.92      4.84

# it also works if x is a string
get_means(test_data, "trad_n")
#> # A tibble: 1 × 6
#>    mean    sd     n std.error conf.low conf.high
#>   <dbl> <dbl> <dbl>     <dbl>    <dbl>     <dbl>
#> 1  4.38  3.66   250      0.23     3.92      4.84

# Let's do that again but add weights
get_means(test_data, trad_n, wt = wts)
#> # A tibble: 1 × 6
#>    mean    sd     n std.error conf.low conf.high
#>   <dbl> <dbl> <dbl>     <dbl>    <dbl>     <dbl>
#> 1  4.39  3.66   250      0.23     3.93      4.84

# the wt argument can also be in quotes like this
get_means(test_data, "trad_n", wt = "wts")
#> # A tibble: 1 × 6
#>    mean    sd     n std.error conf.low conf.high
#>   <dbl> <dbl> <dbl>     <dbl>    <dbl>     <dbl>
#> 1  4.39  3.66   250      0.23     3.93      4.84

# Now let's do the average score for different education levels
get_means(test_data, trad_n, edu_f, wts)
#> # A tibble: 4 × 7
#>   edu_f                mean    sd     n std.error conf.low conf.high
#>   <fct>               <dbl> <dbl> <dbl>     <dbl>    <dbl>     <dbl>
#> 1 High School or Less  5.31  3.68    76      0.42     4.47      6.15
#> 2 Some College         4.15  3.6     79      0.4      3.34      4.96
#> 3 Bachelor's Degree    3.63  3.37    57      0.45     2.73      4.52
#> 4 Graduate Degree      3.63  3.71    38      0.6      2.42      4.85

# it also works with quotes
get_means(test_data, "trad_n", "edu_f", "wts")
#> # A tibble: 4 × 7
#>   edu_f                mean    sd     n std.error conf.low conf.high
#>   <fct>               <dbl> <dbl> <dbl>     <dbl>    <dbl>     <dbl>
#> 1 High School or Less  5.31  3.68    76      0.42     4.47      6.15
#> 2 Some College         4.15  3.6     79      0.4      3.34      4.96
#> 3 Bachelor's Degree    3.63  3.37    57      0.45     2.73      4.52
#> 4 Graduate Degree      3.63  3.71    38      0.6      2.42      4.85

# you can also pipe in the `df` argument if you want to do some data
# transformations before you calculate the means. For example, say you want
# to compare the means of `trad_n` among people who agreed vs disagreed with
# the variable `top`:
test_data %>%
  mutate(top_f2 = make_dicho(top)) %>%
  get_means(trad_n, top_f2, wts)
#> # A tibble: 2 × 7
#>   top_f2    mean    sd     n std.error conf.low conf.high
#>   <fct>    <dbl> <dbl> <dbl>     <dbl>    <dbl>     <dbl>
#> 1 Agree     5.72  3.92    97      0.4      4.93      6.51
#> 2 Disagree  3.52  3.26   153      0.26     3         4.04