Make dichotomous factors

Convert a vector of class factor or haven_labelled to a "dichotomous factor vector". When I refer to a "dichotomous factor vector", I am referring to a vector of class factor with only two levels that are opposites (e.g., "Agree" and "Disagree")

Usage

make_dicho(x, flip_levels = FALSE)

Arguments

x

A vector of type haven_labelled or factor.

flip_levels

Logical. If FALSE, the default, the factor levels are kept the same. If TRUE, the factor levels are flipped. Only specify this if you want to change the order of the factor level.

Details

make_dicho was designed to work on any vector that is of class factor, haven_labelled, or numeric with value labels. If the vector is numeric with no value labels, the function will return an error. This is because the function first converts the vector to a factor using as_factor() haven::as_factor() from the haven package. Then, it removes the first word if there are multiple words in the factor level.

The resulting factor levels default to alphabetical but if you want to reverse them, just set flip_levels = TRUE in the function.

In addition, this function adds two new attributes. The first attribute, transformation, indicates the data transformation that the original vector underwent to create this new vector. The second attribute, label, contains the variable label that was found in the original variable. However, if the original vector did not have a variable label, then this attribute will not show up. This is only useful if you care about

Examples

library(tibble)
library(dplyr)
library(labelled)
library(haven)

# create fake data
df <- tibble::tribble(
  ~x, ~y, ~z,
  3, 2, 3,
  4, 4, 2,
  2, 6, 1,
  1, 1, 4,
  5, 4, 3,
  6, 5, 6
) %>%
# add value labels
labelled::set_value_labels(
  x = c(`Strongly agree` = 1,
        `Agree` = 2,
        `Somewhat agree` = 3,
        `Somewhat disagree` = 4,
        `Disagree` = 5,
        `Strongly disagree` = 6),
  y = c(`Strongly agree` = 1,
        `Agree` = 2,
        `Somewhat agree` = 3,
        `Somewhat disagree` = 4,
        `Disagree` = 5,
        `Strongly disagree` = 6),
  z = c(`Strongly agree` = 1,
        `Agree` = 2,
        `Somewhat agree` = 3,
        `Somewhat disagree` = 4,
        `Disagree` = 5,
        `Strongly disagree` = 6)
) %>%
# add variable labels
labelled::set_variable_labels(
  x = "This is the variable label for x",
  y = "This is the variable label for y",
  z = "This is the variable label for z"
)

# show the data transformation with a haven_labelled vector
dicho_df <- df %>% dplyr::mutate(dicho_x = make_dicho(x))
# check the updated dataset
dicho_df
#> # A tibble: 6 × 4
#>   x                     y                     z                     dicho_x 
#>   <dbl+lbl>             <dbl+lbl>             <dbl+lbl>             <fct>   
#> 1 3 [Somewhat agree]    2 [Agree]             3 [Somewhat agree]    Agree   
#> 2 4 [Somewhat disagree] 4 [Somewhat disagree] 2 [Agree]             Disagree
#> 3 2 [Agree]             6 [Strongly disagree] 1 [Strongly agree]    Agree   
#> 4 1 [Strongly agree]    1 [Strongly agree]    4 [Somewhat disagree] Agree   
#> 5 5 [Disagree]          4 [Somewhat disagree] 3 [Somewhat agree]    Disagree
#> 6 6 [Strongly disagree] 5 [Disagree]          6 [Strongly disagree] Disagree

# Check the attributes
attributes(dicho_df$dicho_x)
#> $levels
#> [1] "Agree"    "Disagree"
#> 
#> $class
#> [1] "factor"
#> 
#> $transformation
#> Converting 'x' to a dichotomous factor with 'Agree' as the reference level
#> 
#> $label
#> [1] "This is the variable label for x"
#> 
# another way of checking the attributes
str(dicho_df$dicho_x)
#>  Factor w/ 2 levels "Agree","Disagree": 1 2 1 1 2 2
#>  - attr(*, "transformation")= 'glue' chr "Converting 'x' to a dichotomous factor with 'Agree' as the reference level"
#>  - attr(*, "label")= chr "This is the variable label for x"

# check the factor levels
unique(dicho_df$dicho_x)
#> [1] Agree    Disagree
#> Levels: Agree Disagree

# ----------------------------------------------------------------------------

# function also works with factors
dicho_df <- df %>%
  dplyr::mutate(
    # convert variable to a factor
    factor_x = haven::as_factor(x),
    # convert the factor to a dichotomous factor
    dicho_x = make_dicho(factor_x)
  )

# check the updated dataset
dicho_df
#> # A tibble: 6 × 5
#>   x                     y                     z                 factor_x dicho_x
#>   <dbl+lbl>             <dbl+lbl>             <dbl+lbl>         <fct>    <fct>  
#> 1 3 [Somewhat agree]    2 [Agree]             3 [Somewhat agre… Somewha… Agree  
#> 2 4 [Somewhat disagree] 4 [Somewhat disagree] 2 [Agree]         Somewha… Disagr…
#> 3 2 [Agree]             6 [Strongly disagree] 1 [Strongly agre… Agree    Agree  
#> 4 1 [Strongly agree]    1 [Strongly agree]    4 [Somewhat disa… Strongl… Agree  
#> 5 5 [Disagree]          4 [Somewhat disagree] 3 [Somewhat agre… Disagree Disagr…
#> 6 6 [Strongly disagree] 5 [Disagree]          6 [Strongly disa… Strongl… Disagr…

# Check the attributes
attributes(dicho_df$dicho_x)
#> $levels
#> [1] "Agree"    "Disagree"
#> 
#> $class
#> [1] "factor"
#> 
#> $transformation
#> Converting 'factor_x' to a dichotomous factor with 'Agree' as the reference level
#> 
#> $label
#> [1] "This is the variable label for x"
#> 
# another way of checking the attributes
str(dicho_df$dicho_x)
#>  Factor w/ 2 levels "Agree","Disagree": 1 2 1 1 2 2
#>  - attr(*, "transformation")= 'glue' chr "Converting 'factor_x' to a dichotomous factor with 'Agree' as the reference level"
#>  - attr(*, "label")= chr "This is the variable label for x"

# check the factor levels
unique(dicho_df$dicho_x)
#> [1] Agree    Disagree
#> Levels: Agree Disagree

# ----------------------------------------------------------------------------
# function also works inside dplyr::across()

# Create new columns using `across()`
dicho_df <- df %>%
  dplyr::mutate(
    # use this example if you don't want to flip the factor levels
    dplyr::across(
      x:z,
      make_dicho,
      .names = "dicho_{col}"
    ),
    # if you want to flip the factor levels, follow this example
    dplyr::across(
      x:z,
      ~make_dicho(., flip_levels = TRUE),
      .names = "dicho_flipped_{col}"
    )
  )
# show that the function worked properly by creating two new sets of variables
dicho_df
#> # A tibble: 6 × 9
#>   x                     y        z       dicho_x dicho_y dicho_z dicho_flipped_x
#>   <dbl+lbl>             <dbl+lb> <dbl+l> <fct>   <fct>   <fct>   <fct>          
#> 1 3 [Somewhat agree]    2 [Agre… 3 [Som… Agree   Agree   Agree   Agree          
#> 2 4 [Somewhat disagree] 4 [Some… 2 [Agr… Disagr… Disagr… Agree   Disagree       
#> 3 2 [Agree]             6 [Stro… 1 [Str… Agree   Disagr… Agree   Agree          
#> 4 1 [Strongly agree]    1 [Stro… 4 [Som… Agree   Agree   Disagr… Agree          
#> 5 5 [Disagree]          4 [Some… 3 [Som… Disagr… Disagr… Agree   Disagree       
#> 6 6 [Strongly disagree] 5 [Disa… 6 [Str… Disagr… Disagr… Disagr… Disagree       
#> # ℹ 2 more variables: dicho_flipped_y <fct>, dicho_flipped_z <fct>

# show the underlying structure of the entire df
str(dicho_df)
#> tibble [6 × 9] (S3: tbl_df/tbl/data.frame)
#>  $ x              : dbl+lbl [1:6] 3, 4, 2, 1, 5, 6
#>    ..@ labels: Named num [1:6] 1 2 3 4 5 6
#>    .. ..- attr(*, "names")= chr [1:6] "Strongly agree" "Agree" "Somewhat agree" "Somewhat disagree" ...
#>    ..@ label : chr "This is the variable label for x"
#>  $ y              : dbl+lbl [1:6] 2, 4, 6, 1, 4, 5
#>    ..@ labels: Named num [1:6] 1 2 3 4 5 6
#>    .. ..- attr(*, "names")= chr [1:6] "Strongly agree" "Agree" "Somewhat agree" "Somewhat disagree" ...
#>    ..@ label : chr "This is the variable label for y"
#>  $ z              : dbl+lbl [1:6] 3, 2, 1, 4, 3, 6
#>    ..@ labels: Named num [1:6] 1 2 3 4 5 6
#>    .. ..- attr(*, "names")= chr [1:6] "Strongly agree" "Agree" "Somewhat agree" "Somewhat disagree" ...
#>    ..@ label : chr "This is the variable label for z"
#>  $ dicho_x        : Factor w/ 2 levels "Agree","Disagree": 1 2 1 1 2 2
#>   ..- attr(*, "transformation")= 'glue' chr "Converting 'x' to a dichotomous factor with 'Agree' as the reference level"
#>   ..- attr(*, "label")= chr "This is the variable label for x"
#>  $ dicho_y        : Factor w/ 2 levels "Agree","Disagree": 1 2 2 1 2 2
#>   ..- attr(*, "transformation")= 'glue' chr "Converting 'y' to a dichotomous factor with 'Agree' as the reference level"
#>   ..- attr(*, "label")= chr "This is the variable label for y"
#>  $ dicho_z        : Factor w/ 2 levels "Agree","Disagree": 1 1 1 2 1 2
#>   ..- attr(*, "transformation")= 'glue' chr "Converting 'z' to a dichotomous factor with 'Agree' as the reference level"
#>   ..- attr(*, "label")= chr "This is the variable label for z"
#>  $ dicho_flipped_x: Factor w/ 2 levels "Disagree","Agree": 2 1 2 2 1 1
#>   ..- attr(*, "transformation")= 'glue' chr "Converting 'x' to a dichotomous factor and reordering the factor levels so that 'Disagree' is the reference level"
#>   ..- attr(*, "label")= chr "This is the variable label for x"
#>  $ dicho_flipped_y: Factor w/ 2 levels "Disagree","Agree": 2 1 1 2 1 1
#>   ..- attr(*, "transformation")= 'glue' chr "Converting 'y' to a dichotomous factor and reordering the factor levels so that 'Disagree' is the reference level"
#>   ..- attr(*, "label")= chr "This is the variable label for y"
#>  $ dicho_flipped_z: Factor w/ 2 levels "Disagree","Agree": 2 2 2 1 2 1
#>   ..- attr(*, "transformation")= 'glue' chr "Converting 'z' to a dichotomous factor and reordering the factor levels so that 'Disagree' is the reference level"
#>   ..- attr(*, "label")= chr "This is the variable label for z"

# show how the levels are flipped when "flip_levels = TRUE"
levels(dicho_df$dicho_x)
#> [1] "Agree"    "Disagree"
levels(dicho_df$dicho_flipped_x)
#> [1] "Disagree" "Agree"