Default theme

This function creates the default theme that all ADL theme functions are built off of. It functions similarly to theme_gray, in that all of the default ggplot2 themes are built off theme_gray.

Usage

theme_default(
  base_size = 12,
  base_family = "L",
  base_line_size = base_size/24,
  base_rect_size = base_size/24,
  legend_position = "right",
  axis_text = TRUE,
  axis_text_x = TRUE,
  axis_text_y = TRUE,
  grid = TRUE,
  grid_x_only = FALSE,
  grid_y_only = FALSE,
  grid_major = TRUE,
  grid_minor = TRUE,
  grid_major_x = TRUE,
  grid_major_y = TRUE,
  grid_minor_x = TRUE,
  grid_minor_y = TRUE
)

Arguments

base_size

Base font size, given in pts. Also controls the spacing in the graph.

base_family

Base font family. Default is Lato.

base_line_size

Base size for line elements.

base_rect_size

Base size for rect elements.

legend_position

The position of the legend. Options are: "left", "right", "top", "bottom", or "none". "none" removes the legend. "right" is the default.

axis_text

Logical. Determines if BOTH axes have labels. If TRUE, the default, both axes are labelled. If FALSE, neither axis is labelled. Note, this controls both axes. If you want to remove only one axis, use the axis_text_x or axis_text_y.

axis_text_x

Logical. Determines if the x-axis has labels. If TRUE, the default, the x-axis labels are shown. If FALSE, the x-axis labels are removed from the plot.

axis_text_y

Logical. Determines if the y-axis has labels. If TRUE, the default, the y-axis labels are shown. If FALSE, the y-axis labels are removed from the plot.

grid

Logical. Determines if ALL grid lines should appear. If TRUE, the default, all grid lines appear. If FALSE, all grid lines disappear.

grid_x_only

Logical. Determines if only x-axis grid lines (vertical lines) should appear. If FALSE, the default, all grid lines appear. If TRUE, only the x-axis grid lines appear and the y-axis grid lines will disappear.

grid_y_only

Logical. Determines if only y-axis grid lines (horizontal lines) should appear. If FALSE, the default, all grid lines appear. If TRUE, only the y-axis grid lines appear and the x-axis grid lines will disppear.

grid_major

Logical. Determines if the major grid lines should appear. If TRUE, the default, the major grid lines will appear. If FALSE, the major grid lines will disappear.

grid_minor

Logical. Determines if the minor grid lines should appear. If TRUE, the default, the minor grid lines will appear. If FALSE, the minor grid lines will disappear.

grid_major_x

Logical. Determines if the major x-axis grid lines will appear. If TRUE, the default, the major x-axis grid lines will appear. If FALSE, the major x-axis grid lines will disappear.

grid_major_y

Logical. Determines if the major y-axis grid lines will appear. If TRUE, the default, the major y-axis grid lines will appear. If FALSE, the major t-axis grid lines will disappear.

grid_minor_x

Logical. Determines if the minor x-axis grid lines will appear. If TRUE, the default, the minor x-axis grid lines will appear. If FALSE, the minor x-axis grid lines will disappear.

grid_minor_y

Logical. Determines if the minor y-axis grid lines will appear. If TRUE, the default, the minor y-axis grid lines will appear. If FALSE, the minor t-axis grid lines will disappear.